Move/rename nav menu module.
