+++ components/scoreboard/ScoreboardShareButton.vue
@@ -2,7 +2,7 @@
 import type { ButtonPassThroughOptions } from 'primevue/button';
 import { useToast } from "primevue/usetoast";
 
-import { scoreboardMenuItem } from '~~/nav';
+import { scoreboardMenuItem } from '~~/data/nav';
 import type { ID, Team } from '~~/types';
 
 const toast = useToast();
